+++ MrLocalBackend/Services/Helpers/ValidateData.cs
@@ -17,7 +17,7 @@ namespace MrLocalBackend.Services.Helpers
         {
             _shopRepository = shopRepository;
         }
-        public async Task<bool> ValidateProductData(string shopId, string name, string description, double? price, bool isUpdate, string priceType)
+        public async Task<bool> ValidateProductData(string shopId, string name, string description, decimal? price, bool isUpdate, string priceType)
         {
             static bool IsStringEmpty(string str) => str == null || str.Length == 0;
